refactor(nix): autostart xdg applications
This commit refactors the `xdg.autostart` configuration in `users/yechiel/home/home.nix`. It introduces a `let` binding to define a shorthand `e` for `pkgs.lib.getExe`, making the list of autostart entries cleaner and more maintainable. The previous string-based paths are replaced with calls to this new `e` function for `bitwarden`, `trayscale`, `beeper`, and `element-desktop`.